Move relative functions into BaseAction.
