+++ balancer-js/src/modules/swaps/joinExit/actions/helpers.ts
@@ -1,107 +1,5 @@
-import { BigNumber } from '@ethersproject/bignumber';
-import { Relayer, OutputReference } from '@/modules/relayer/relayer.module';
-import { subSlippage } from '@/lib/utils/slippageHelper';
-import { ActionStep, ActionType, Actions } from './types';
+import { ActionType, Actions } from './types';
 import { Swap } from './swap';
-
-/**
- * If its not the first action then the amount will come from the previous output ref
- * @param amount
- * @param actionType
- * @param actionStep
- * @param opRefKey
- * @returns
- */
-export function getActionAmount(
-  amount: string,
-  actionType: ActionType,
-  actionStep: ActionStep,
-  opRefKey: number
-): string {
-  let amountIn = amount;
-  if (
-    actionStep === ActionStep.TokenOut ||
-    (actionStep === ActionStep.Middle && actionType === ActionType.Join) ||
-    (actionStep === ActionStep.Middle && actionType === ActionType.Exit)
-  ) {
-    amountIn = Relayer.toChainedReference(opRefKey - 1).toString();
-  }
-  return amountIn;
-}
-
-function getOutputRef(key: number, index: number): OutputReference {
-  const keyRef = Relayer.toChainedReference(key);
-  return { index: index, key: keyRef };
-}
-
-/**
- * If its not the final action then we need an outputReferece to chain to next action as input
- * @param actionStep
- * @param tokenOutIndex
- * @param opRefKey
- * @returns
- */
-export function getActionOutputRef(
-  actionStep: ActionStep,
-  tokenOutIndex: number,
-  opRefKey: number
-): [OutputReference, number] {
-  let opRef: OutputReference = {} as OutputReference;
-  if (actionStep === ActionStep.TokenIn || actionStep === ActionStep.Middle) {
-    opRef = getOutputRef(opRefKey, tokenOutIndex);
-    opRefKey++;
-  }
-  return [opRef, opRefKey];
-}
-
-/**
- * Use slippage to set min amount out
- * @param amountOut
- * @param slippage
- * @returns
- */
-export function getActionMinOut(amountOut: string, slippage: string): string {
-  // Currently only handle ExactIn swap. ExactOut would add slippage
-  // We should apply slippage to each swaps amountOut
-  return subSlippage(
-    BigNumber.from(amountOut),
-    BigNumber.from(slippage)
-  ).toString();
-}
-
-/**
- * Find if the Action is:
- * Direct: tokenIn > tokenOut
- * TokenIn: tokenIn > chain...
- * TokenOut: ...chain > tokenOut
- * Middle: ...chain > action > chain...
- * @param tokenInIndex
- * @param tokenOutIndex
- * @param tokenInIndexAction
- * @param tokenOutIndexAction
- * @returns
- */
-export function getActionStep(
-  tokenInIndex: number,
-  tokenOutIndex: number,
-  tokenInIndexAction: number,
-  tokenOutIndexAction: number
-): ActionStep {
-  let actionStep: ActionStep;
-  if (
-    tokenInIndexAction === tokenInIndex &&
-    tokenOutIndexAction === tokenOutIndex
-  ) {
-    actionStep = ActionStep.Direct;
-  } else if (tokenInIndexAction === tokenInIndex) {
-    actionStep = ActionStep.TokenIn;
-  } else if (tokenOutIndexAction === tokenOutIndex) {
-    actionStep = ActionStep.TokenOut;
-  } else {
-    actionStep = ActionStep.Middle;
-  }
-  return actionStep;
-}
 
 /**
  * Find the number of actions that end with tokenOut
